+++ packages/service-library/requirements/_base.txt
@@ -4,28 +4,23 @@
 #
 #    pip-compile --output-file=_base.txt _base.in
 #
-aiohttp-session[secure]==2.7.0
 aiohttp==3.5.4
 aiopg[sa]==0.16.0
-asn1crypto==0.24.0        # via cryptography
 async-timeout==3.0.1      # via aiohttp
-attrs==19.1.0             # via aiohttp, openapi-core
-cffi==1.12.3              # via cryptography
+attrs==19.1.0             # via aiohttp, jsonschema, openapi-core
 chardet==3.0.4            # via aiohttp
-cryptography==2.6.1
 idna-ssl==1.1.0           # via aiohttp
 idna==2.8                 # via idna-ssl, yarl
-jsonschema==2.6.0
+jsonschema==3.0.1
 lazy-object-proxy==1.4.1  # via openapi-core
 multidict==4.5.2          # via aiohttp, yarl
 openapi-core==0.8.0
-openapi-spec-validator==0.2.6  # via openapi-core
-pathlib==1.0.1            # via openapi-spec-validator
+openapi-spec-validator==0.2.7  # via openapi-core
 psycopg2-binary==2.8.2
-pycparser==2.19           # via cffi
+pyrsistent==0.15.2        # via jsonschema
 pyyaml==5.1
-six==1.12.0               # via cryptography, openapi-core, openapi-spec-validator
-sqlalchemy[postgresql_psycopg2binary]==1.3.3
+six==1.12.0               # via jsonschema, openapi-core, openapi-spec-validator, pyrsistent
+sqlalchemy[postgresql_psycopg2binary]==1.3.4
 strict-rfc3339==0.7       # via openapi-core
 typing-extensions==3.7.2  # via aiohttp
 ujson==1.35
