+++ packages/service-library/requirements/_test.txt
@@ -4,51 +4,48 @@
 #
 #    pip-compile --output-file=_test.txt _test.in
 #
-aiohttp-session[secure]==2.7.0
 aiohttp==3.5.4
 aiopg[sa]==0.16.0
-asn1crypto==0.24.0
 astroid==2.2.5            # via pylint
 async-timeout==3.0.1
 atomicwrites==1.3.0       # via pytest
 attrs==19.1.0
 certifi==2019.3.9         # via requests
-cffi==1.12.3
 chardet==3.0.4
 coverage==4.5.3           # via coveralls, pytest-cov
 coveralls==1.7.0
-cryptography==2.6.1
 docopt==0.6.2             # via coveralls
 idna-ssl==1.1.0
 idna==2.8
+importlib-metadata==0.15  # via pluggy
 isort==4.3.20             # via pylint
-jsonschema==2.6.0
+jsonschema==3.0.1
 lazy-object-proxy==1.4.1
 mccabe==0.6.1             # via pylint
 more-itertools==7.0.0     # via pytest
 multidict==4.5.2
 openapi-core==0.8.0
-openapi-spec-validator==0.2.6
-pathlib==1.0.1
-pluggy==0.11.0            # via pytest
+openapi-spec-validator==0.2.7
+pluggy==0.12.0            # via pytest
 psycopg2-binary==2.8.2
 py==1.8.0                 # via pytest
-pycparser==2.19
 pylint==2.3.1
+pyrsistent==0.15.2
 pytest-aiohttp==0.3.0
 pytest-cov==2.7.1
-pytest-runner==4.4
+pytest-runner==5.1
 pytest==4.5.0
 pyyaml==5.1
 requests==2.22.0          # via coveralls
 six==1.12.0
-sqlalchemy[postgresql_psycopg2binary]==1.3.3
+sqlalchemy[postgresql_psycopg2binary]==1.3.4
 strict-rfc3339==0.7
 typed-ast==1.3.5          # via astroid
 typing-extensions==3.7.2
 ujson==1.35
-urllib3==1.25.2           # via requests
+urllib3==1.25.3           # via requests
 wcwidth==0.1.7            # via pytest
 werkzeug==0.15.4
 wrapt==1.11.1             # via astroid
 yarl==1.3.0
+zipp==0.5.1               # via importlib-metadata
