+++ services/web/server/requirements/_base.txt
@@ -8,18 +8,22 @@ aadict==0.2.3             # via asset
 aio-pika==2.9.0
 aiohttp-jinja2==1.1.1
 aiohttp-security==0.4.0
+aiohttp-session[secure]==2.7.0
 aiohttp==3.5.4
 aiopg[sa]==0.16.0
 aiosmtplib==1.0.5
 amqp==2.4.2               # via kombu
+asn1crypto==0.24.0        # via cryptography
 asset==0.6.12             # via passwordmeter
 async-timeout==3.0.1      # via aiohttp
 asyncpg==0.18.3
 attrs==19.1.0             # via aiohttp, openapi-core
 billiard==3.6.0.0         # via celery
 celery==4.3.0
+cffi==1.12.3              # via cryptography
 change-case==0.5.2
 chardet==3.0.4            # via aiohttp
+cryptography==2.6.1
 globre==0.1.5             # via asset
 idna-ssl==1.1.0           # via aiohttp
 idna==2.8                 # via idna-ssl, yarl
@@ -32,20 +36,20 @@ lazy-object-proxy==1.4.1  # via openapi-core
 markupsafe==1.1.1         # via jinja2
 multidict==4.5.2          # via aiohttp, yarl
 openapi-core==0.8.0
-openapi-spec-validator==0.2.6  # via openapi-core
+openapi-spec-validator==0.2.7  # via openapi-core
 passlib==1.7.1
 passwordmeter==0.1.8
-pathlib==1.0.1            # via openapi-spec-validator
 pika==0.10.0              # via aio-pika
 psycopg2-binary==2.8.2
-python-engineio==3.5.1    # via python-socketio
-python-socketio==4.0.1
+pycparser==2.19           # via cffi
+python-engineio==3.7.0    # via python-socketio
+python-socketio==4.0.3
 pytz==2019.1              # via celery
 pyyaml==5.1
 semantic-version==2.6.0
 shortuuid==0.5.0          # via aio-pika
-six==1.12.0               # via aadict, asset, openapi-core, openapi-spec-validator, python-engineio, python-socketio, tenacity
-sqlalchemy[postgresql_psycopg2binary]==1.3.3
+six==1.12.0               # via aadict, asset, cryptography, openapi-core, openapi-spec-validator, python-engineio, python-socketio, tenacity
+sqlalchemy[postgresql_psycopg2binary]==1.3.4
 strict-rfc3339==0.7       # via openapi-core
 tenacity==5.0.4
 trafaret-config==2.0.2
