+++ services/web/server/src/simcore_service_webserver/config/host-dev-config.yaml
@@ -55,3 +55,5 @@ projects:
 storage:
   host: storage
   port: 11111
+session:
+  secret_key: "TODO: Replace with a key of at least length 32"
