+++ services/web/server/src/simcore_service_webserver/config/server-defaults.yaml
@@ -54,3 +54,5 @@ rest:
     - http://localhost:9081
 projects:
   location: http://localhost:8043/api/specs/webserver/v0/components/schemas/project-v0.0.1.json
+session:
+  secret_key: "TODO: Replace with a key of at least length 32"
