+++ services/web/server/src/simcore_service_webserver/config/server-docker-dev.yaml
@@ -55,4 +55,6 @@ storage:
   host: ${STORAGE_HOST}
   port: ${STORAGE_PORT}
   version: v0
+session:
+  secret_key: "TODO: Replace with a key of at least length 32"
 ...
