+++ services/web/server/src/simcore_service_webserver/config/server-template.yaml
@@ -52,4 +52,6 @@ rest:
   location: http://${APIHUB_HOST}:${APIHUB_PORT}/api/specs/webserver/v0/openapi.yaml
 projects:
   location: http://${APIHUB_HOST}:${APIHUB_PORT}/api/specs/webserver/v0/components/schemas/project-v0.0.1.json
+session:
+  secret_key: "Thirty  two  length  bytes  key."
 ...
