+++ services/web/server/src/simcore_service_webserver/session_config.py
@@ -1,0 +1,12 @@
+""" session subsystem's configuration
+
+    - config-file schema
+    - settings
+"""
+import trafaret as T
+
+CONFIG_SECTION_NAME = 'session'
+
+schema = T.Dict({
+    "secret_key": T.String
+})
