+++ services/web/server/tests/unit/with_postgres/config.yaml
@@ -51,3 +51,5 @@ rest:
 projects:
   enabled: False
   location: ${OSPARC_SIMCORE_REPO_ROOTDIR}/api/specs/webserver/v0/components/schemas/project-v0.0.1.json
+session:
+  secret_key: "TODO: Replace Thirty  two  length  bytes  key."
